requirements.txt
boto3==1.34.19
botocore==1.34.19
certifi==2023.11.17
charset-normalizer==3.3.2
filelock==3.13.1
fsspec==2023.12.2
huggingface-hub==0.20.2
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.0
numpy==1.26.3
packaging==23.2
Pillow==9.3.0
PyMySQL==1.1.0
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
s3transfer==0.10.0
safetensors==0.4.1
six==1.16.0
sympy==1.12
tokenizers==0.15.0
torch==2.1.2+cpu
torchaudio==2.1.2+cpu
torchvision==0.16.2+cpu
tqdm==4.66.1
transformers==4.36.2
typing_extensions==4.9.0
urllib3==2.0.7
